@@ -10,7 +10,7 @@ export default {
   components: {
     LearnMode
   },
-  data() {
+  setup() {
     return {
       terms: [
         {
